BJS2-10425
- создал класс со статическими методами
- убрал комменты;
- задачу перенес в другой пакет
