+++ src/main/java/faang/school/godbless/stream/streamAPI_1/Main.java
@@ -1,0 +1,9 @@
+package faang.school.godbless.stream.streamAPI_1;
+
+
+public class Main {
+
+
+    public static void main(String[] args) {
+    }
+}
